After adding BallGame and fixing all compiler warnings

+++ GameEngine/GameEngine/IOManager.cpp
@@ -14,11 +14,11 @@ namespace GameEngine {
 
 		file.seekg(0, std::ios::end);
 
-		int fileSize = file.tellg();
+		int fileSize = (int)file.tellg();
 
 		file.seekg(0, std::ios::beg);
 
-		fileSize -= file.tellg();
+		fileSize -= (int)file.tellg();
 
 		buffer.resize(fileSize);
